완료 -JWT로 로그인 -로그인한 사용자의 CRUD 권한 부여 -Guest의 조회 권한 부여
수정 필요
- 로그인 토큰을 전달한 채로 로그인 API 또는 회원가입 API를 호출한 경우
- 댓글 목록
= 로그인 유저의 아이디 표시

+++ src/main/java/com/sparta/board/controller/PostController.java
@@ -3,7 +3,10 @@ package com.sparta.board.controller;
 import com.sparta.board.model.Post;
 import com.sparta.board.repository.PostRepository;
 import com.sparta.board.dto.PostRequestDto;
+import com.sparta.board.security.UserDetailsImpl;
+import io.jsonwebtoken.io.IOException;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,19 +19,20 @@ public class PostController {
     private final PostRepository postRepository;
 
     @PostMapping("/api/posts")
-    public Post createPost(@RequestBody PostRequestDto requestDto){
-        Post post = new Post(requestDto);
-        return postRepository.save(post);
-    }
-
-    @GetMapping("/api/posts")
-    public List<Post> getPosts(){
-        return postRepository.findAll();
+    public Post createPost(@RequestBody PostRequestDto requestDto, @AuthenticationPrincipal UserDetailsImpl userDetails){
+            String writer = userDetails.getUsername();
+            Post post = new Post(requestDto, writer);
+            return postRepository.save(post);
     }
 
     @DeleteMapping("/api/posts/{id}")
-    public Long deletePost(@PathVariable Long id){
-        postRepository.deleteById(id);
+    public Long deletePost(@PathVariable Long id, @AuthenticationPrincipal UserDetailsImpl userDetails){
+        String username = userDetails.getUsername();
+        Post post = postRepository.findById(id).orElseThrow(
+                () -> new IllegalArgumentException("존재하지 않는 계정입니다.")
+        );
+        if(post.getWriter().equals(username)) postRepository.deleteById(id);
+        else throw new IllegalArgumentException("계정이 일치하지 않습니다.");
         return id;
     }
